server/backend/requirements.txt
fastapi
uvicorn
pydantic
python-socketio[asgi]
langchain-openai
langgraph
pandas
openpyxl